feat: add Board tests and implement Board.vue with session filtering, navigation, and UI enhancements
